backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0
django-celery-results==2.5.1
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==25.1
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flower==2.0.1
gunicorn==23.0.0
hiredis==3.1.0
humanize==4.12.3
kombu==5.5.3
Markdown==3.8
packaging==25.0
pillow==11.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tornado==6.4.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
